# integration-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CodeSentinel X API
  version: 1.0.0
  description: Enterprise code security scanner API (REST contract)
servers:
  - url: https://api.codesentinelx.example.com
paths:
...
  /v1/scans:
    post:
      summary: Start a full or differential scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScanRequest'
      responses:
        "202":
          description: Scan accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanJob'
  /v1/scans/{scan_id}:
    get:
      summary: Get scan status and summary
      parameters:
        - in: path
          name: scan_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanSummary'
...
components:
  schemas:
...
    CreateScanRequest:
      type: object
      properties:
        project_id: { type: string }
        repository_url: { type: string }
        branch: { type: string }
        commit_sha: { type: string }
        mode: { type: string, enum: [full, differential] }
      required: [project_id, mode]
    ScanJob:
      type: object
      properties:
        scan_id: { type: string }
        status: { type: string, enum: [queued, running, completed, failed] }
        queued_at: { type: string, format: date-time }
    ScanSummary:
      type: object
      properties:
        scan_id: { type: string }
        status: { type: string }
        risk_score: { type: number }
        severity_distribution:
          type: object
          properties:
            Critical: { type: integer }
            High: { type: integer }
            Medium: { type: integer }
            Low: { type: integer }
            Info: { type: integer }
...